requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
jinja2==3.1.2
python-multipart==0.0.6
aiosmtplib==2.0.2
email-validator==2.0.0
httpx==0.24.1
bs4==0.0.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0 